添加经纬度筛选排序逻辑

+++ forum-api/src/main/java/pub/developers/forum/api/request/article/ArticleAdminPageRequest.java
@@ -46,6 +46,6 @@ public class ArticleAdminPageRequest implements Serializable {
     private BigDecimal latitude;
 
     @ApiModelProperty("半径")
-    private Integer radius;
+    private BigDecimal radius;
 
 }
